scripts/requirements-pxart-extract.txt
numpy>=1.26,<3
scipy>=1.11,<2
pillow>=10,<13
